pyproject.toml
[project]
name = "climate_guard"
version = "0.1.0"
description = "A web scraper for Latvian news articles"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
dependencies = [
    "gradio>=4.44.1",
    "python-dotenv>=1.0.1",
    # Jupyter
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    # Scrapping
    "requests>=2.26.0",
    "beautifulsoup4>=4.10.0",
    "newspaper3k>=0.2.8",
    "lxml[html-clean]>=5.3.0",
    "gdeltdoc>=1.5.0",
    "lxml-html-clean>=0.2.2",
    # Pandas parquet support
    "pyarrow>=18.0.0",
    # LLMs
    "tiktoken>=0.8.0",
    "groq>=0.11.0",
    "openai>=1.51.0",
    "anthropic>=0.34.2",
    "litellm>=1.51.2",
    "boto3>=1.35.52",
    # Classification metrics
    "scikit-learn>=1.5.2",
    # ROUGE score
    "evaluate>=0.4.3",
    "absl-py>=2.1.0",
    "rouge-score>=0.1.2",
    "deepmultilingualpunctuation>=1.0.1",
    # Plotting
    "plotly>=5.24.1",
    "kaleido>=0.2.1",
    "seaborn>=0.13.2",
    "datasets>=2.14.4",
    "streamlit>=1.40.0",
    "streamlit-aggrid>=1.0.5",
]
requires-python = ">=3.12"

# https://github.com/astral-sh/uv/issues/7703#issuecomment-2376674577
[tool.uv]
constraint-dependencies = ["kaleido!=0.2.1.post1"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["climate_guard"]


[tool.black]
line-length = 100
target-version = ['py312']
